pyproject.toml
[project]
name = "odl-salesforce-agent"
version = "0.1.0"
description = "Eval suite for Casey, the ODL Salesforce intake agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agentforce-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
casey-eval = "eval.cli:cli"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["eval"]
